Reworked the implementation.
